+++ tsk/fs/fatfs.c
@@ -522,7 +522,7 @@ fatfs_block_walk(TSK_FS_INFO * fs, TSK_DADDR_T a_start_blk,
                 cnt =
                     tsk_fs_read_block(fs, addr, data_buf,
                     fs->block_size * 8);
-                if (cnt != fs->block_size * 8) {
+                if (cnt != (ssize_t)(fs->block_size * 8)) {
                     if (cnt >= 0) {
                         tsk_error_reset();
                         tsk_error_set_errno(TSK_ERR_FS_READ);
